Added mission token authentication for certain mission endpoints
Mission model now has a mission token that can be used as static authentication (when no regular auth can be provided). This currently allows for embedding of mission details and slotlist on non-public missions e.g. into a website or forum
Added endpoints for retrieving, generating and deleting the mission token
Added migration for `missionToken` column
Closes #80